Fully cleaned the project.

+++ Source/Rule56/Gui/Core/GUIUtility.Text.cs
@@ -1,10 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using RimWorld;
 using UnityEngine;
 using Verse;
-using Verse.Sound;
 using GUITextState = System.Tuple<string, CombatAI.Gui.GUIFontSize, System.Tuple<float, float>, System.Tuple<int, int, int, int>, System.Tuple<UnityEngine.FontStyle, UnityEngine.FontStyle, UnityEngine.FontStyle, UnityEngine.FontStyle>>;
 
 namespace CombatAI.Gui
@@ -31,7 +28,7 @@ namespace CombatAI.Gui
 			{
 				return text;
 			}
-			return text.Substring(0, (int)((float)text.Length * height / rect.height)) + "...";
+			return text.Substring(0, (int)(text.Length * height / rect.height)) + "...";
 		}
 
 		public static float GetTextHeight(this string text, Rect rect)
